move on contacts change to subscribers dir
